Make test blog posts passing markdown linting
This removes a significant amount of the styling examples, but I think
this is fine as it now only contains styling examples of markdown we are
permitting through the linting. If someone wants to do something not
permitted by the linting they will complain and we can alter the config.
For now I've tried to leave the config as default as possible (online
line length extended for matching python line length), as if we are
using `black` and thus avoiding as much discussion on formatting as
possible.